requirements.txt
gin==0.1.6
gin_config==0.5.0
h5py==3.8.0
imageio==2.31.2
matplotlib==3.5.3
numpy==1.21.5
opencv_python==4.8.1.78
pandas==1.3.5
Pillow==11.0.0
PyYAML==6.0.2
scikit_image==0.19.3
scipy==1.7.3
seaborn==0.13.2
timm==0.9.8
tqdm==4.66.1
yacs==0.1.8
cython

